+++ webapp/package.json
@@ -57,7 +57,6 @@
     "jest": "26.1.0",
     "jest-canvas-mock": "2.2.0",
     "jest-junit": "11.0.1",
-    "mattermost-webapp": "github:mattermost/mattermost-webapp#0b27dbfaa79b88cc207d835bbe604577d9397425",
     "node-sass": "4.14.1",
     "sass-loader": "9.0.2",
     "style-loader": "1.2.1",
@@ -65,9 +64,12 @@
     "webpack-cli": "3.3.12"
   },
   "dependencies": {
+    "@testing-library/jest-dom": "5.11.9",
+    "@types/debounce": "1.2.0",
     "core-js": "3.6.5",
     "debounce": "1.2.1",
-    "mattermost-redux": "5.27.0",
+    "mattermost-redux": "5.28.1",
+    "mattermost-webapp": "github:mattermost/mattermost-webapp#0b27dbfaa79b88cc207d835bbe604577d9397425",
     "react": "16.13.1",
     "react-redux": "7.2.0",
     "redux": "4.0.5",
